use middleware to reduce code duplication
